ckb-debugger-tests: generate_sighash_all move to lib.rs
